+++ NEWS.md
@@ -1,6 +1,8 @@
 ## 19.5.0 (in-progress)
 
+* Status date set on newly created instance records (MODINVSTOR-509)
 * Item status date is set for newly created item records (MODINVSTOR-508)
+
 
 ## 19.4.0 2020-10-08